% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_gaps.R
\name{base_gaps}
\alias{base_gaps}
\title{Return a list of the indices of blocks of sequential \code{NAs} in \code{base}.}
\usage{
base_gaps(base)
}
\arguments{
\item{base}{baseline(s) generated by \code{\link{sus}}.}
}
\value{
A list of the indices of blocks of sequential \code{NAs} in
 \code{base}. If \code{base} has no \code{NAs}, return \code{NULL}. If a
 block of \code{NAs} exists before the first sustained shift it is ignored.
}
\description{
Calling \code{\link{sus}} returns a vector \code{base} with some values
\code{NA} and some not. \code{base_gaps()} returns a list of the indices of
these blocks of \code{NAs}. This function is used by \code{\link{multi_shift}}.
}
\examples{
base_gaps(c(rep(NA,5), rep(3,8), rep(NA,25)))
}
\seealso{
\code{\link{ticker}} \code{\link{multi_shift}}
}

