% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_basic.R
\name{basic_shift}
\alias{basic_shift}
\title{Iterate through a vector and find the indices of all shifts relative to a
single (constant) baseline.}
\usage{
basic_shift(base, val, trigger = 6)
}
\arguments{
\item{base}{A numeric vector of length one. The current baseline value.}

\item{val}{A numeric vector. The vector \code{basic_shift} iterates over.}

\item{trigger}{The default number of observations for a shift is 6.}
}
\value{
If shifts are found, the indices of these shifts. If not, \code{NULL}.
}
\description{
Call this function once to iterate through all elements of a vector and return
the indices of all shifts relative to a single (constant) baseline. For a
changing baseline this function is used multiple times.
}
\examples{
runchart:::basic_shift(base = 2, val = 1:8)
runchart:::basic_shift(base = 2, val = 1:7)
}
\seealso{
\code{\link{multi_shift}}
}
\keyword{internal}

