% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_basic.R
\name{basic_trend}
\alias{basic_trend}
\title{Iterate through a vector and find the indices of all trends.}
\usage{
basic_trend(val, trigger = 5)
}
\arguments{
\item{val}{A numeric vector. The vector \code{basic_trend} iterates over.}

\item{trigger}{The default number of observations for a trend is 5.}
}
\value{
If trends are found, the indices of these trends. If not, \code{NULL}.
}
\description{
Call this function once to iterate through all elements of a vector and return
the indices of all trends relative to a single (constant) baseline. For a
changing baseline this function is used multiple times by \code{multi_trend()}.
}
\examples{
runchart:::basic_trend(val = 1:8)
runchart:::basic_trend(val = 1:7)
}
\seealso{
\code{\link{basic_shift}}
}
\keyword{internal}

