% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df}
\alias{df}
\title{Create a \code{data.frame} from the vectors \code{base} and \code{base_ext}.}
\usage{
df(base, base_ext, base_label = rep(NA_character_, length(base)))
}
\arguments{
\item{base}{A numeric vector.}

\item{base_ext}{A numeric vector.}

\item{base_label}{A character vector.}
}
\value{
A \code{data.frame} with 2 columns: \code{base} and \code{base_ext}.
}
\description{
Create a \code{data.frame} from the vectors \code{base} and \code{base_ext}.
}
\examples{
runchart:::df(c(rep(4.5, 8), NA), rep(4.5, 9))
}
