% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elems8}
\alias{elems8}
\title{First n non-NA elements of x where n defaults to 8.}
\usage{
elems8(x, n = 8L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{How many non-NA elements of x should be returned?}
}
\value{
A numeric vector.
}
\description{
Creating the first baseline of a run chart requires the first 8 non-NA data
points. \code{elems8(x, n)} finds the first n non-NA elements of x
where n defaults to 8.
}
\examples{
runchart:::elems8(1:9)
runchart:::elems8(1:9, 5)
}
\seealso{
\code{\link{index1}}, \code{\link{index8}}
}
\keyword{internal}
