% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Convenience infix function.}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A logical vector.
}
\description{
Convenience infix function.
}
\examples{
runchart:::`\%notin\%`(1:3, 3:5)
}
