% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index1}
\alias{index1}
\title{Index of the first non-NA element of x.}
\usage{
index1(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of length one.
}
\description{
Index of the first non-NA element of x.
}
\examples{
runchart:::index1(c(NA,1:9))
}
\seealso{
\code{\link{elems8}}, \code{\link{index8}}
}
