% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index8}
\alias{index8}
\title{Index of the nth non-NA element of x where n defaults to 8.}
\usage{
index8(x, n = 8L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{Which non-NA element to return}
}
\value{
A numeric vector of length one.
}
\description{
Index of the nth non-NA element of x where n defaults to 8.
}
\examples{
runchart:::index8(c(NA,1:9))
}
\seealso{
\code{\link{elems8}}, \code{\link{index1}}
}
