% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fields.R
\name{rc_fields}
\alias{rc_fields}
\title{Create run chart fields for vector \code{val}.}
\usage{
rc_fields(val)
}
\arguments{
\item{val}{create run chart fields for \code{val}.}
}
\value{
A data.frame with four columns: 1. The original data. 2. A base line.
 3. An extended base line. 4: Shifts.
}
\description{
Create run chart fields for vector \code{val}.
}
\examples{
runchart:::rc_fields(rep(1:20))
}
\seealso{
\code{\link{sus}} \code{\link{multi_shift}}
}
\keyword{internal}

