% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_plot.R
\name{rc_plot}
\alias{rc_plot}
\title{Plot a run chart.}
\usage{
rc_plot(df)
}
\arguments{
\item{df}{a data.frame for \code{rc_plot()} to plot. \code{df} must consist of
exactly two columns: a column of type Date called \code{date} and a column of
numeric type called \code{value}.}
}
\value{
A plot.
}
\description{
Plot a run chart.
}
\examples{
n <- 30
runchart:::rc_plot(
data.frame(
 date = seq.Date(Sys.Date(), by = "week", length.out = n),
 value = cumsum(rnorm(n))))
}
\seealso{
\code{\link{rc_fields}} \code{\link{split}}
}
\keyword{internal}
