% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rebase}
\alias{rebase}
\title{Rebase \code{base} starting from \code{start} until \code{start} using the
median of \code{new_vals}}
\usage{
rebase(base, start = 1, end = length(base), new_vals)
}
\arguments{
\item{base}{A numeric vector.}

\item{start}{Where to rebase \code{base} from (inclusive).}

\item{end}{Where to rebase \code{base} until (inclusive).}

\item{new_vals}{Rebase \code{base} using the median of \code{new_vals}}
}
\value{
A numeric vector.
}
\description{
Rebase \code{base} starting from \code{start} until \code{start} using the
median of \code{new_vals}
}
\examples{
runchart:::rebase(1:9, start = 4, new_vals = 1)
runchart:::rebase(1:9, start = 4, new_vals = c(1,2,3))
}
