% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchart.R
\docType{package}
\name{runchart-package}
\alias{runchart}
\alias{runchart-package}
\title{runchart}
\description{
This is a package for creating run charts.
}
\details{
The runchart package allows users to create run charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsphdms/runchart}
  \item Report bugs at \url{https://github.com/jsphdms/runchart/issues}
}

}
\author{
\strong{Maintainer}: Joseph Adams \email{joseph@josephadams.net}

}
