% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchart.R
\name{runchart}
\alias{runchart}
\title{Create run chart fields for vector \code{val}.}
\usage{
runchart(val)
}
\arguments{
\item{val}{create run chart fields for \code{val}.}
}
\value{
A data.frame with four columns: 1. The original data. 2. A base line.
 3. An extended base line. 4: Shifts.
}
\description{
Create run chart fields for vector \code{val}.
}
\examples{
runchart(rep(1:20))
}
\seealso{
\code{\link{sus}} \code{\link{shift}}
}

