% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Rephased baselines are joined up when plotted as a single vector. This means
that a line will join the end of a base line to the beginning of another.
Visually, this isn't very useful. In order to avoid this, we split baselines
into dataframes with a column for each rephased value. \code{split} is a
utility function which splits a vector into a dataframe for this purpose.}
\usage{
split(v, vname)
}
\arguments{
\item{v}{a vector for \code{split} to split into a dataframe.}

\item{vname}{the columns of the returned dataframe are named \code{vname1},
\code{vname2}, \code{vname3}, etc}
}
\value{
A data.frame with a column for each new value of \code{v}.
}
\description{
Rephased baselines are joined up when plotted as a single vector. This means
that a line will join the end of a base line to the beginning of another.
Visually, this isn't very useful. In order to avoid this, we split baselines
into dataframes with a column for each rephased value. \code{split} is a
utility function which splits a vector into a dataframe for this purpose.
}
\examples{
runchart:::split(
v = c(1,1,1,1,1,1,5,5,5,5,5,5,5,5,3,3,3,3,3,3,3),
vname = "example"
)
}
\seealso{
\code{\link{sus}} \code{\link{multi_shift}}
}
\keyword{internal}

