% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus.R
\name{sus}
\alias{sus}
\title{Create baseline(s) for \code{val}.}
\usage{
sus(val, trigger = 9)
}
\arguments{
\item{val}{\code{sus()} iterates along the elements of \code{val} to create
baselines.}

\item{trigger}{Number of data points required to create a sustained shift.}
}
\value{
A data.frame with two columns: 1. The median of the first n useful
 observations of a sustained shift (where n defaults to 8) 2: The same as the
 previous column, but with each median value extended to either the end of
 \code{val} or the beginning of the next sustained shift.
}
\description{
The first baseline for \code{val} is the median of the first 8 useful
observations of \code{val}. It stretches from the first useful observation to
the 8th useful observation and is often black. A second line of the same value
is usually extended from the beginning of the black line to the end of
\code{val}. This second line is for visual comparison later in \code{val} and
is usually grey.
}
\details{
If a sustained shift is detected after the first 8 useful data points, the
baseline is redrawn using the first 8 useful observations in the sustained
shift. Given a vector \code{val} the function \code{sus()} returns both the
grey and black lines in a \code{data.frame}.
}
\examples{
runchart:::sus(rep(1:20), trigger = 9)
}
\seealso{
\code{\link{ticker}}
}
\keyword{internal}

