% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_first.R
\name{sus_first}
\alias{sus_first}
\title{Indices of the first sustained shift of \code{val}.}
\usage{
sus_first(base, val, trigger = 9)
}
\arguments{
\item{base}{A numeric vector of length \code{1}.}

\item{val}{\code{sus_first} iterates along the elements of \code{val}.}

\item{trigger}{Number of data points required to create a sustained shift.}
}
\value{
The indices of the first sustained shift of \code{val}. \code{sus_first}
 returns \code{NULL} if no sustained shift is found.
}
\description{
A sustained shift is defined by 9 consecutive data points either all above or
all below the current baseline.
}
\examples{
runchart:::sus_first(base = 0, val = 1:20)
}
\seealso{
\code{\link{ticker}} \code{\link{sus}}
}
\keyword{internal}

