% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Iterate through a vector and find the first sustained shift.}
\usage{
ticker(base, val, trigger = 9)
}
\arguments{
\item{base}{A numeric vector of length one. The current baseline value.}

\item{val}{A numeric vector. The vector \code{ticker} iterates over.}

\item{start}{The index of the first useful element of \code{val}. Begin
iterating here.}

\item{updown}{Takes value \code{1} or \code{-1}. Is the first useful element
of \code{val} above (\code{1}) or below (\code{-1}) \code{base}?}
}
\value{
\code{NULL}, until either a sustained shift is found (then it returns
 the indices of the first sustained shift) or no sustained shift is found
 (then it returns the string \code{No sus shift found}). The indices of the
 first sustained shift (if one is found). Otherwise, \code{NULL}.
}
\description{
Use this function to iterate through a vector by calling it multiple times
until it returns either: the indices of the first sustained shift; or the
string \code{No sus shift found}.
}
\examples{
tick <- ticker(base = 2, val = 1:15, start = 1, updown = -1)
tick()
tick()
tick()
etc
}
\seealso{
\code{\link{sus}}
}

