% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_multi.R
\name{multi_shift}
\alias{multi_shift}
\title{Create shift(s) for \code{val} using baselines generated by
\code{\link{sus}}.}
\usage{
multi_shift(val, base, base_ext, trigger = 6)
}
\arguments{
\item{val}{\code{multi_shift()} iterates along the elements of \code{val} to
create shifts.}

\item{base}{baseline generated by \code{\link{sus}}.}

\item{base_ext}{extended baseline generated by \code{\link{sus}}.}

\item{trigger}{Number of data points required to create a shift.}
}
\value{
A vector with all shifts observed in \code{val}. If no shifts observed,
 a vector full of \code{NA}s.
}
\description{
Shifts occur when 6 or more consecutive useful observations (non-NA and
different from the current baseline) land all above/below the baseline.
}
\examples{
multi_shift(rep(1:20), bases = c(rep(4.5,8), rep(NA,12)), trigger = 9)
}
\seealso{
\code{\link{basic_shift}}
}

