% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Iterate through a vector and find the first sustained shift.}
\usage{
ticker(base, val, trigger = 9)
}
\arguments{
\item{base}{A numeric vector of length one. The current baseline value.}

\item{val}{A numeric vector. The vector \code{ticker} iterates over.}

\item{trigger}{A numeric vector of length one. Specifies how many consecutive
points are required for a sustained shift.}
}
\value{
\code{NULL}, until either a sustained shift is found (then it returns
 the indices of the first sustained shift) or no sustained shift is found
 (then it returns the string \code{No sus shift found}). The indices of the
 first sustained shift (if one is found). Otherwise, \code{NULL}.
}
\description{
Use this function to iterate through a vector by calling it multiple times
until it returns either: the indices of the first sustained shift; or the
string \code{No sus shift found}.
}
\examples{
tick <- runchart:::ticker(base = 2, val = 1:15)
tick()
tick()
tick()
}
\seealso{
\code{\link{sus}}
}
\keyword{internal}

